fix: migration for project release

+++ packages/server/api/src/app/database/migration/postgres/1734886495968-AddMappingForProject.ts
@@ -8,26 +8,34 @@ export class AddMappingForProject1734886495968 implements MigrationInterface {
             ALTER TABLE "project"
             ADD "mapping" jsonb
         `)
-        await queryRunner.query(`
-            UPDATE "project"
-            SET "mapping" = "git_repo"."mapping"
-            FROM "git_repo"
-            WHERE "project"."id" = "git_repo"."projectId"
-        `)
-        await queryRunner.query(`
-            ALTER TABLE "git_repo" DROP COLUMN "mapping"
-        `)
 
+        const hasGitRepoTable = await queryRunner.hasTable('git_repo')
+        if (!hasGitRepoTable) {
+            await queryRunner.query(`
+                UPDATE "project"
+                SET "mapping" = "git_repo"."mapping"
+                FROM "git_repo"
+                WHERE "project"."id" = "git_repo"."projectId"
+            `)
+            await queryRunner.query(`
+                ALTER TABLE "git_repo" DROP COLUMN "mapping"
+            `)
+        }
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
+        const hasGitRepoTable = await queryRunner.hasTable('git_repo')
+        
         await queryRunner.query(`
             ALTER TABLE "project" DROP COLUMN "mapping"
         `)
-        await queryRunner.query(`
-            ALTER TABLE "git_repo"
-            ADD "mapping" jsonb
-        `)
+
+        if (hasGitRepoTable) {
+            await queryRunner.query(`
+                ALTER TABLE "git_repo"
+                ADD "mapping" jsonb
+            `)
+        }
     }
 
 }
